+++ CHANGES.rst
@@ -43,6 +43,9 @@ Imviz
   that are of a different pixel scale or are rotated w.r.t. the reference data when
   linked by WCS. [#2154]
 
+- Fixed wrong angle translations between sky regions in ``regions`` and ``photutils``.
+  They were previously off by 90 degrees. [#2154]
+
 Mosviz
 ^^^^^^